feat: Refactor testing utilities and add web endpoint tests with snapshots
