+++ nicknamer/server/tests/common/mod.rs
@@ -1,7 +1,82 @@
+use axum::http::StatusCode;
 use migration::MigratorTrait;
 use sea_orm::{Database, DatabaseConnection};
+use serde::Serialize;
+use std::collections::BTreeMap;
 use testcontainers_modules::testcontainers::runners::AsyncRunner;
 use testcontainers_modules::{postgres, testcontainers};
+
+/// Headers that vary between test runs and should be filtered out for stable snapshots.
+pub const VARIABLE_HEADERS: &[&str] = &[
+    "date",
+    "expires",
+    "last-modified",
+    "etag",
+    "server",
+    "x-request-id",
+    "x-trace-id",
+    "set-cookie",
+    "content-length",
+];
+
+/// HTTP response snapshot for testing endpoints.
+#[derive(Debug, Serialize)]
+pub struct HttpResponseSnapshot {
+    pub test_context: String,
+    pub status: u16,
+    pub headers: BTreeMap<String, String>,
+    pub html_body: Vec<String>,
+}
+
+impl HttpResponseSnapshot {
+    /// Create a new HTTP response snapshot.
+    pub fn new(
+        body_text: &str,
+        status: StatusCode,
+        headers: &axum::http::HeaderMap,
+        test_context: &str,
+    ) -> Self {
+        Self {
+            test_context: test_context.to_string(),
+            status: status.as_u16(),
+            headers: filter_variable_headers(headers),
+            html_body: normalize_html_for_snapshot(body_text),
+        }
+    }
+
+    /// Create a snapshot from just HTML content and context.
+    /// This is useful for testing template rendering without HTTP responses.
+    pub fn from_html(html_content: &str, test_context: &str) -> Self {
+        Self {
+            test_context: test_context.to_string(),
+            status: 200,
+            headers: BTreeMap::new(),
+            html_body: normalize_html_for_snapshot(html_content),
+        }
+    }
+}
+
+/// Normalize HTML content for consistent snapshots by removing dynamic values.
+pub fn normalize_html_for_snapshot(html: &str) -> Vec<String> {
+    // Split HTML by newlines and convert to Vec<String>
+    // In the future, we could add more sophisticated normalization
+    html.lines().map(|line| line.to_string()).collect()
+}
+
+/// Filter out variable headers from response headers for snapshot testing.
+pub fn filter_variable_headers(headers: &axum::http::HeaderMap) -> BTreeMap<String, String> {
+    headers
+        .iter()
+        .filter_map(|(name, value)| {
+            let name_str = name.as_str().to_lowercase();
+            if VARIABLE_HEADERS.contains(&name_str.as_str()) {
+                None
+            } else {
+                value.to_str().ok().map(|v| (name_str, v.to_string()))
+            }
+        })
+        .collect()
+}
 
 pub async fn setup_container() -> anyhow::Result<testcontainers::ContainerAsync<postgres::Postgres>>
 {
